iox-#615 fixed clang mac os compile error

+++ iceoryx_posh/include/iceoryx_posh/internal/popo/building_blocks/chunk_distributor.inl
@@ -166,7 +166,7 @@ inline void ChunkDistributor<ChunkDistributorDataType>::deliverToAllStoredQueues
                                               remainingQueues.begin(),
                                               remainingQueues.end(),
                                               queueIntersection.begin());
-            queueIntersection.resize(iter - queueIntersection.begin());
+            queueIntersection.resize(static_cast<uint64_t>(iter - queueIntersection.begin()));
             remainingQueues = queueIntersection;
 
             // deliver to remaining queues
